src/infra/router.py
from rest_framework import routers
from app.users.serial import *

# Routers provide an easy way of automatically determining the URL conf.
router = routers.DefaultRouter()
router.register(r'users', UserViewSet)