requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-filter==22.1
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
pycodestyle==2.9.1
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
